fixes in user interface and update readme
