+++ src/modules/graphics/UserInterface.js
@@ -1,299 +1,6 @@
 import Vector from '../physics/Vector.js';
-import {colouredText, Data, assert} from '../helpers.js';
-
-export class Terminal extends Data {
-  // TODO: bind data to terminal
-  constructor(shared) {
-    super(shared);
-
-    this.sharedData.highlighted = false;
-
-    this.commandMap = new Map(
-      [
-        ["focus", this.focusCommand],
-        ["unfocus", this.unfocus],
-        ["centre", this.highlightCommand],
-        ["move", this.moveCommand],
-        ["sample", this.setSampleSizeCommand]
-      ]
-    );
-    this.output = "";
-    this.input = "";
-    this.element = document.createElement("div");
-  }
-
-  focusCommand = (args) => {
-
-    if(args.length > 0)
-    {
-      // check input is int
-      const selection = args[0];
-      if (!isNaN(selection) && (parseFloat(selection) | 0) === parseFloat(selection))
-      {
-        if (parseInt(selection)<=(this.sharedData.nodes.length -1))
-        {
-          const datapoint = this.sharedData.nodes[parseInt(selection)];
-          if(datapoint.stroke=="red")
-          {
-            datapoint.stroke="black";
-            this.sharedData.highlighted=false;
-          }
-          else
-          {
-            datapoint.stroke = "red";
-            if(this.sharedData.highlighted!==false)
-            {
-              this.sharedData.nodes[this.sharedData.highlighted].stroke = "black";
-            }
-            this.sharedData.highlighted=parseInt(selection);
-            this.log("selected node "+colouredText("#"+selection,"green"))
-          }
-        }
-        else
-        {
-          this.logError("invalid range");
-        }
-      }
-      else
-      {
-        this.logError("input is not an integer");
-      }
-    }
-    else if (this.sharedData.highlighted!==false)
-    {
-      this.log("focused on existing highlighed node.")
-    }
-    else
-    {
-      this.logError("no user input or highlighted node.")
-      return;
-    }
-
-    //TODO: neighbours are now stored in forces, check for springs...
-    this.sharedData.nodes.forEach((node) => {
-      const neighbours = this.sharedData.nodes[this.sharedData.highlighted].neighbours;
-      let isNeighbour = false;
-      debugger;
-      for(let i = 0; i < neighbours.length; i++)
-      {
-        if(node.id === neighbours[i][0])
-        {
-          isNeighbour = true
-        }
-      }
-
-      if(node.id != highlighted)
-      {
-        this.sharedData.nodes[node.id].showEdges = false;
-
-        if(!isNeighbour){
-          node.visible = false;
-        }
-      }
-
-    });
-
-
-  }
-
-  unfocus(){
-    data.forEach((node) => {
-        node.showEdges = true;
-        node.visible = true;
-        node.edgeStroke = "black";
-    });
-
-    log("focus off")
-  }
-
-  updateScroll(){
-    this.element.scrollTop = this.element.scrollHeight;
-  }
-
-  initTerminal()
-  {
-
-    this.element.addEventListener("focus", (event) =>{
-      this.element.style.color = "black";
-      this.element.innerHTML = this.output + "UserIn: " + this.input + "<";
-    })
-
-    this.element.addEventListener("focusout", (event) => {
-      this.element.style.color = "rgb(173,172,173)";
-    });
-
-    this.element.addEventListener("keydown", (event) => {
-
-      const key = event.keyCode;
-      const char = String.fromCharCode((96 <= key && key <= 105) ? key-48 : key).toLowerCase();
-
-      // If the user has pressed enter
-
-      switch (key) {
-        case 32 /*space*/:
-        {
-          this.input+=(this.input.slice(-1)===" "?"":" ");
-          this.element.innerHTML = this.output + "UserIn: " + this.input + "<";;
-          break;
-        }
-        case 13 /*enter*/:
-        {
-          const args = this.input.split(" ");
-          const command = args[0]
-          if(!this.commandMap.has(command))
-          {
-            this.input = "";
-            logError("no such command.")
-            break;
-          } else {
-            const fn = this.commandMap.get(command);
-            fn(args.slice(1));
-            this.input = "";
-            this.element.innerHTML = this.output + "UserIn: " + this.input + "<";
-            break;
-          }
-        }
-        case 8/*backspace*/:
-        {
-          if(this.input.length>0)
-          {
-            this.input = this.input.slice(0, -1);
-            this.element.innerHTML = this.output + "UserIn: " + this.input + "<";
-          }
-          break;
-        }
-        case 37/*left*/:
-        {
-
-          break;
-        }
-        case 39/*right*/:
-        {
-
-          break;
-        }
-          case 38/*up*/:
-        {
-
-          break;
-        }
-          case 40/*down*/:
-
-            break;
-        default:
-        {
-          this.input += char;
-          this.element.innerHTML = this.output + "UserIn: " + this.input + "<";
-        }
-
-      }
-    }, false);
-  }
-
-  clearTerminal(){
-    output = input = "";
-  }
-
-  logDebug (msg, newline=true){
-    this.output += colouredText("WeLD (debug): ","blue") + msg + (newline?"<br/>":"");
-    this.element.innerHTML = this.output + "UserIn: " + this.input + "<";
-    this.updateScroll()
-}
-
-  logWarning (msg, newline=true){
-    this.output += colouredText("WeLD (warning): ","orange") + msg + (newline?"<br/>":"");
-    this.element.innerHTML = this.output + "UserIn: " + this.input + "<";
-    this.updateScroll()
-  }
-
-
-  logError (msg, newline=true){
-    this.output += colouredText("WeLD (error): ","red") + msg + (newline?"<br/>":"");
-    this.element.innerHTML = this.output + "UserIn: " + this.input + "<";
-    this.updateScroll()
-  }
-
-
-  log (msg, newline=true) {
-    this.output += colouredText("WeLD: ","green") + msg + (newline?"<br/>":"") ;
-    this.element.innerHTML = this.output + "UserIn: " + this.input + "<";
-    this.updateScroll()
-  };
-
-  setSampleSizeCommand = (args) => {
-    // TODO formalise commands
-    var size = args[0];
-    if( !isNaN(size) && (parseFloat(size) | 0) === parseFloat(size))
-    {
-      size = parseInt(size)
-      if(this.sharedData.sampleSize > size){
-        this.sharedData.samples1 = this.sharedData.samples1.slice(-size, -1);
-        this.sharedData.samples2 = this.sharedData.samples2.slice(-size, -1);
-        this.sharedData.samples3 = this.sharedData.samples3.slice(-size, -1);
-      }
-      this.sharedData.sampleSize = size
-      this.log(`sample size is now ${size}`);
-    }
-    else
-    {
-      this.logError("input is not an integer");
-    }
-  }
-
-  highlightCommand(args){
-    const selection = args[0];
-    if( !isNaN(selection) && (parseFloat(selection) | 0) === parseFloat(selection))
-    {
-      if(parseInt(selection)<=(data.length -1))
-      {
-        //highlighted = selection;
-        this.focus(selection);
-        this.log("centred node #"+selection);
-      } else {
-        this.logError("invalid range");
-      }
-    } else {
-            this.logError("input is not an integer");
-          }
-  }
-
-  // move a node to the specified location
-  // for debugging purposes
-  moveCommand = (args) =>
-  {
-    var [nodeID, x, y, z] = args;
-    let relative = false;
-
-    if (args.length >= 5 && args[4] == "r") {
-      relative = true;
-    }
-
-    if (nodeID && x && y && z) {
-      nodeID = parseInt(nodeID)
-      x = parseFloat(x);
-      y = parseFloat(y);
-      z = parseFloat(z);
-
-      const node = this.sharedData.nodes[nodeID];
-
-      if (relative) {
-        node.ri.x += x;
-        node.ri.y += y;
-        node.ri.z += z;
-      }
-      else {
-        node.ri.x = x;
-        node.ri.y = y;
-        node.ri.z = z;
-      }
-
-    } else {
-      logError("Usage: move [ID] [x] [y] [z] [r (optional)], e.g. move 0 -13 45 -279");
-    }
-
-  }
-
-}
+import { Data } from '../helpers.js';
+import { Terminal } from './Terminal.js';
 
 export class UserInterface extends Data
 {
@@ -425,17 +132,15 @@ export class UserInterface extends Data
     // initialise terminal after styling;
     this.terminal.initTerminal();
 
-
-
     this.infoBox = document.createElement("div");
     control.appendChild(this.infoBox);
     this.infoBox.setAttribute("id", "infoBox");
     this.infoBox.style.position = "static";
     this.infoBox.style.padding = "2.5px";
-    this.infoBox.style.backgroundColor = "rgba(0,0,0,0.7)";
+    this.infoBox.style.backgroundColor = "rgba(0,0,0,0.2)";
     this.infoBox.style.width = "100%";
     this.infoBox.style.height = "33.333%";
-    this.infoBox.style.color = "rgb(173,172,173)";
+    this.infoBox.style.color = "black"//"rgb(173,172,173)";
     this.infoBox.style.zIndex = document.getElementById("control").style.zIndex + 1;
     this.infoBox.style.overflowX = "hidden";
     this.infoBox.style.overflowY = "scroll";
@@ -578,10 +283,8 @@ export class UserInterface extends Data
     context.strokeStyle = col;
 
     context.moveTo( GRAPH_LEFT, ( GRAPH_HEIGHT - dataArr[ 0 ] / largest * GRAPH_HEIGHT ) + GRAPH_TOP );
-    // draw reference value for day of the week
     for( var i = 1; i < arrayLen; i++ ){
         context.lineTo( GRAPH_RIGHT / arrayLen * i + GRAPH_LEFT, ( GRAPH_HEIGHT - dataArr[ i ] / largest * GRAPH_HEIGHT ) + GRAPH_TOP );
-        // draw reference value for day of the week
     }
 
     context.stroke();
