+++ src/namespaces/Graphics.js
@@ -1,4 +1,5 @@
-import {UserInterface, Terminal} from '../modules/graphics/UserInterface.js';
+import { UserInterface } from '../modules/graphics/UserInterface.js';
+import { Terminal } from '../modules/graphics/Terminal.js';
 import { Renderer } from '../modules/graphics/Renderer.js';
 
 const Graphics = {UserInterface, Terminal, Renderer};
